final touch: rename to RemoveUselessReadOnlyTagRector
