--- rules-tests/DeadCode/Rector/Property/RemoveUselessReadOnlyDocRector/config/configured_rule.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\DeadCode\Rector\Property\RemoveUselessReadOnlyDocRector;
-
-return RectorConfig::configure()
-    ->withRules([RemoveUselessReadOnlyDocRector::class]);
