+++ rules-tests/DeadCode/Rector/Property/RemoveUselessReadOnlyTagRector/Fixture/skip_with_description.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\Property\RemoveUselessReadOnlyDocRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\Property\RemoveUselessReadOnlyTagRector\Fixture;
 
 class SkipHasDescription
  {
