+++ rules-tests/DeadCode/Rector/Property/RemoveUselessReadOnlyTagRector/config/configured_rule.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\Config\RectorConfig;
+use Rector\DeadCode\Rector\Property\RemoveUselessReadOnlyTagRector;
+
+return RectorConfig::configure()
+    ->withRules([RemoveUselessReadOnlyTagRector::class]);
